Implement post plot callback for theming
Some theming relevant settings can't be set via rcParams but must
be applied to all ax in a plot. Add the possibility to submit post
plot callbacks, which get called with each axis in a plotting
figure.

The theme_grey submits such a callback.

Also change how the default theming is applied, so that it is
possible to print a plot with the default style and then add the theme
without getting the callback/rcParams from the default theme applied.

What is not possible anymore is adding theme_grey (with a callback) and
then adding theme_matplotlib (currently without a callback) without
having some "leftover" from theme_grey (the callback).