Implemented scale_facet_wrap function.
Added support for calculating the limits and setting the label visibility for
four different scale types when the facet type is "wrap": None, "free",
"free_x", and "free_y". The behavior of the different scale types is based on
the behavior outlined in the ggplot2 docs
(http://docs.ggplot2.org/current/facet_wrap.html), except I decided to display
the x-axis labels on the bottom plot for each column, not just the plots in
the bottom row.