Use out own image testing framework for anti-aliased text
Currently baseline images, which were generated on *ix(?)
systems fail on windows computer as windows systems produces
different baseline images which are *NOT* anti-aliased.
I traced that to a line in matplotlib, which setups the
rc context to not use anti aliasing and it seems that the
rc-context used during `ggplot.draw()` which uses
anti-aliased fonts is not used when the image is
saved in the `image_comparison` decorator.

This implements replicates the matplotlib testing framework, but
sets anti-aliased to true and also uses only pngs.

There are now two ways to test for same image:

* `@image_comparison` decorator (similar as in matplotlib)
* `assert_same_images()`