Use our own image testing framework for anti-aliased text
Currently baseline images, which were generated on *ix(?)
systems fail on windows computer as windows systems produces
different baseline images which are *NOT* anti-aliased.
I traced that to a line in matplotlib, which setups the
rc context to not use anti aliasing and it seems that the
rc-context used during `ggplot.draw()` which uses
anti-aliased fonts is not used when the image is
saved in the `image_comparison` decorator.

This commit replicates the matplotlib testing framework, but
sets anti-aliased to true and also uses only pngs.

There are now two ways to test for same image:

* `@image_comparison` decorator (similar as in matplotlib)
* `assert_same_ggplot(gg, "name")`

The latter needs some setup to to locate the baseline images:

```
from . import get_assert_same_ggplot, cleanup
assert_same_ggplot = get_assert_same_ggplot(__file__)
```

This fix is also needed when we want to support matplotlib 1.1.1,
as this old version does not support using the `image_comparison`
decorator outside of matplotlib.