TST: make visual tests more robust
Up to now, when a test produced a matlotlib plot and did not
use the `@cleanup` decorator (`@image_comparison did it internaly)
the plot was available in the next test and could lead to
problems when that test used `@images_comparision`.

No assert that no old plots are available when the testcase is
setup/ before calling the real test function.

Also fix the breakup that made me look into this problem (randomly
failing test cases).

Also convert one testcase to the new `assert_same_ggplot` way
instead of the decorator.