FIX: don't turn off the first axis in some cases
The first axis was turned off when the number of axis was
a square and exactly the number of preallocated axis. The
trigger seems to be that [1,2,3][-0:] will return the whole
list but [1,2,3][-x:] only the last x element (the algo
expected that [..][0] returns an empty list). Why only the
first axis was turned off and not all of them and only
when the number of axis was a sqaure: i don't know :-(