Lines -- slope and intercepts
ggplot2 line geoms seem to allow mapping to parameters!
A clearer view for these "mappable parameters" is to treat them
aesthetics in the geoms (users know they can map to them) and
as parameters in the stats (users discover they can assign functions
which compute on a specific column and yield the required result).