Fixed bug causing legends to print first 3 items only
When using scale_colour_brewer legends would consists only of the first
3 items because of the evaluation order of a python get statement.

The code ```a.get('color', a['fill'])``` would throw a ```KeyError```
when the key `fill` wasn't present even when `color` was present and
there wasn't any reason to evaluate the second argument of `get` per se.

I originally to blame for this bug myself.