Changed the default `edgecolors` parameter to `'none'`
The matplotlib documentation for the `edgecolors` parameter of the
`scatter` method states the following:

> "If None, defaults to (patch.edgecolor). ... If it is ‘none’, the
patch boundary will not be drawn."

The default value is currently `None` which makes all of the points in
a scatter plot show a contrasting edge. This does not match the style
of ggplot2 and has a very "noisy" look. This change switches the
`edgecolors` parameter to `'none'`, which removes the edge completely
makes the plot match ggplot2's look and also reduces visual noise in
the graph.