Correctly sort Categorical facets.
There's a bug where you get an error if you facet a geom_bar with a
single facet variable which is Categorical. This is an old bug, but
it's more visible now that there's a reason to use Categorical facet
variables. I think it's due to a bug in pandas:
https://github.com/pydata/pandas/issues/14011